atualizando sala

+++ src/app/components/sala/sala.component.ts
@@ -7,7 +7,6 @@ import { SalaService } from './../../shared/service/sala.service';
 import { SalaEquipamento, SalaModel, TipoSala } from './../../shared/model/sala.model';
 import { Component, OnInit } from '@angular/core';
 import { MenuItem, MessageService, SelectItem, ConfirmationService } from 'primeng/api';
-import { EquipamentoComponent } from '../equipamento/equipamento.component';
 
 @Component({
   selector: 'app-sala',
@@ -40,7 +39,7 @@ export class SalaComponent implements OnInit {
     equipamentos: null
   };
 
-  
+
 
   ngOnInit(): void {
     this.getAll();
@@ -70,7 +69,16 @@ export class SalaComponent implements OnInit {
   }
 
   confirm(sala: SalaModel) {
-    this.deleteWithButton(sala);
+    this.confirmationService.confirm({
+      message: 'Você tem certeza que deseja excluir esta sala?',
+      accept: () => {
+        this.deleteWithButton(sala);
+      },
+      reject: () => {
+        this.displayEditDialog = false;
+        this.displaySaveDialog = false;
+      }
+    });
   }
   deleteWithButton(sala: SalaModel) {
     this.salaService.delete(sala).subscribe(
@@ -104,11 +112,6 @@ export class SalaComponent implements OnInit {
     });
   }
 
-  getSalaEquipamentoDialog()
-  {
-    
-  }
-
   getAll() {
     this.salaService.getSalas().subscribe(
       (result: any) => {
@@ -135,6 +138,27 @@ export class SalaComponent implements OnInit {
     )
   }
 
+  editEquipWithButton() {
+    this.sala.equipamentos = this.salaEquipamentoService.getSalaEquipamentos();
+    this.salaService.edit(this.sala).subscribe(
+      () => {
+        this.displayEditDialog = false;
+        this.sala = new SalaModel;
+        this.messageService.add({
+          severity: 'success',
+          summary: "Resultado", detail: "Equipamento editado com sucesso"
+        });
+        this.getAll();
+      },
+      error => {
+        this.messageService.add({
+          severity: 'error', summary: "Error",
+          detail: "Equipamento não pode ser editado, verifique os dados e tente novamente"
+        })
+      }
+    )
+  }
+
   getTipoSala(id: number) {
     return TipoSala[id];
   }
@@ -149,5 +173,8 @@ export class SalaComponent implements OnInit {
   }
 
   showEditDialog(sala: SalaModel) {
+    this.displayEditDialog = true;
+    this.sala = sala;
+    this.getAll();
   }
 }
